Trait pallet_mt_rpc_runtime_api::MerkleTreeApi

pub trait MerkleTreeApi<Block: BlockT, E: ElementTrait>: Core<Block> {
    fn get_leaf(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        tree_id: u32, 
        index: u32
    ) -> Result<Option<E>, ApiError> { ... }
    fn get_leaf_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        tree_id: u32, 
        index: u32
    ) -> Result<Option<E>, ApiError> { ... }
}

Provided Methods

fn get_leaf(
&self,
__runtime_api_at_param__: &BlockId<Block>,
tree_id: u32,
index: u32
) -> Result<Option<E>, ApiError>

Get the leaf of tree id at a given index.

fn get_leaf_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
tree_id: u32,
index: u32
) -> Result<Option<E>, ApiError>

Get the leaf of tree id at a given index.